requirements_client.txt
httpx
pydantic
python-dotenv
streamlit